Split into separate CSS and HTML plugins
