+++ packages/astro/src/vite-plugin-build-css/resolve.ts
