+++ packages/astro/src/vite-plugin-build-html/add-rollup-input.ts
