+++ packages/astro/src/vite-plugin-build-html/built-urls.ts
