+++ packages/astro/src/vite-plugin-build-html/extract-assets.ts
